fix: no es necesario dos archivos para esto
